packages/backends/pyproject.toml
[project]
name = "backends"
version = "0.1.0"
description = "Backend connection management for Snowflake and DuckDB"
requires-python = ">=3.11"
dependencies = [
    "schemas",
    "transforms",
    "ibis-framework[duckdb]>=9.0.0",
    "pydantic-settings>=2.0.0",
]

[tool.uv.sources]
schemas = { workspace = true }
transforms = { workspace = true }
